Create new-compiler frontend
