+++ .gitignore
@@ -1,3 +1,9 @@
 *.exe
 build/
 .vscode/
+cmake-build-debug/
+.idea/
+CMakeLists.txt
+test.c
+test.s
+test2.s
